refactor: refactor all project (comments, filenames)
