+++ async-race/src/scripts/App.types.ts
@@ -1,8 +1,1 @@
-// export type RouteType = 'garage' | 'winners';
-
 export type AvailableRoutesType = 'garage' | 'winners';
-
-// export interface IAppState {
-//   availableRoutes: AvailableRoutesType;
-//   activeRoute: RouteType;
-// }
