+++ async-race/src/scripts/views/Garage/Garage.types.ts
@@ -1,0 +1,5 @@
+import { CarType } from '../../API/types';
+
+export interface IGarageState {
+  cars: CarType[];
+}
